+++ src/main/java/znu/visum/ApplicationContext.java
@@ -3,10 +3,12 @@ package znu.visum;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @Configuration
 @ComponentScan
+@EnableScheduling
 public class ApplicationContext {
   @Bean
   public BCryptPasswordEncoder encoder() {