+++ src/main/java/znu/visum/components/movies/domain/models/MovieMetadata.java
@@ -3,15 +3,15 @@ package znu.visum.components.movies.domain.models;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @AllArgsConstructor
 @Builder
 @Getter
 public class MovieMetadata {
 
-  private Long id;
-  private Long movieId;
-  private Long tmdbId;
+  @Setter private Long movieId;
+  @Setter private Long tmdbId;
   private String imdbId;
   private String originalLanguage;
   private String tagline;